# mnemosyne-web-interface/mnemosyne-api/docs/swagger-spec.yaml
openapi: 3.0.0
paths:
...
  /api/company/create-company:
    post:
      operationId: CompanyController_createCompany
      summary: Endpoint is responsible for company creation.
      parameters: []
      requestBody:
        required: true
        description: In order to create a company account user has to provide company
          name, location, website, owner email, list of company members
          (optional) and accept T&C.
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCompanyDto"
      responses:
        "201":
          description: As a response function returns message about created company.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompanyCreatedDto"
        "400":
          description: Bad request is returned in case if company is already exists.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompanyExistsException"
        "403":
          description: Forbidden response for company creation includes only response
            saying that company has not accepted Terms & Conditions.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TacNotAcceptedException"
      tags: &a4
        - Company
      security:
        - basicAuth: []
  /api/company/company-information:
    get:
      operationId: CompanyController_getCompanyInformationById
      summary: Endpoint is responsible for getting information about the company by ID.
      parameters: []
      responses:
        "200":
          description: As a response endpoint returns information about the company by ID.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetCompanyByIdDto"
      tags: *a4
      security:
        - x-access-token: []
        - basicAuth: []
  /api/company/company-public-information:
    get:
      operationId: CompanyController_getCompanyPublicInformation
      summary: Endpoint is responsible for getting company public information.
      parameters:
        - name: companyId
          required: true
          in: query
          description: Company ID
          schema:
            type: string
        - name: page
          required: true
          in: query
          description: Query for page in order to get list of company users.
          schema:
            type: string
        - name: pageSize
          required: true
          in: query
          description: Query for limit in order to get list of company users.
          schema:
            type: string
        - name: query
          required: false
          in: query
          description: Member query
          schema:
            type: string
      responses:
        "200":
          description: As a response user gets all public information about company
            including users.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetCompanyByIdDto"
        "400":
          description: Bad request error is thrown in case of parse exception.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ParseException"
        "404":
          description: Not found error is thrown in case if company ID has been modified
            and company not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompanyNotFoundException"
      tags: *a4
      security:
        - basicAuth: []
  /api/company/company-users:
    get:
      operationId: CompanyController_getCompanyUsers
      summary: Endpoint is responsible for getting the list of all users in company
        with pagination.
      parameters:
        - name: page
          required: true
          in: query
          description: Query for page in order to get list of company users.
          schema:
            type: string
        - name: pageSize
          required: true
          in: query
          description: Query for limit in order to get list of company users.
          schema:
            type: string
        - name: query
          required: false
          in: query
          description: Member query
          schema:
            type: string
      responses:
        "200":
          description: As the response user gets the list of all users along with count
            for pagination.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetCompanyUsersDto"
        "400":
          description: Parse exception is thrown in case if page or limit params cannot be
            parsed as number.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ParseException"
      tags: *a4
      security:
        - x-access-token: []
        - basicAuth: []
  /api/company/update-company:
    patch:
      operationId: CompanyController_updateCompanyInformation
      summary: Endpoint is responsible for updating company information.
      parameters: []
      requestBody:
        required: true
        description: Body contains company information.
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCompanyDto"
      responses:
        "201":
          description: As a response user (with administrator permissions) gets message
            that information has been updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompanyUpdatedDto"
      tags: *a4
      security:
        - x-access-token: []
        - basicAuth: []
  /api/company/transfer-ownership:
    post:
      operationId: CompanyController_transferCompanyOwnership
      summary: Endpoint is responsible for transferring the ownership of the account.
      parameters: []
      requestBody:
        required: true
        description: Body contains new owner email and role id for the old owner along
          with MFA.
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransferOwnershipDto"
      responses:
        "201":
          description: As a response user gets a message that ownership has been
            transferred or anything else related to MFA.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/CompanyOwnershipTransferredDto"
                  - $ref: "#/components/schemas/FullMfaRequiredDto"
                  - $ref: "#/components/schemas/TokenTwoFaRequiredDto"
                  - $ref: "#/components/schemas/PhoneMfaRequiredDto"
        "400":
          description: Bad request error is thrown in case if there is something wrong
            with MFA.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/SmsExpiredException"
                  - $ref: "#/components/schemas/WrongCredentialsException"
                  - $ref: "#/components/schemas/WrongCodeException"
        "404":
          description: Not found error is thrown in case if user has not been found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserNotFoundException"
      tags: *a4
      security:
        - x-access-token: []
        - basicAuth: []
  /api/company/delete-company:
    delete:
      operationId: CompanyController_deleteCompanyAccount
      summary: Endpoint is responsible for deletion of the company account.
      parameters: []
      requestBody:
        required: true
        description: In the body user provides MFA code(s), recovery keys with the
          passphrase, and the password
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeleteCompanyDto"
      responses:
        "201":
          description: As the response endpoint returns MFA requirement and after MFA is
            provided correctly, the message that the company has been deleted.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/CompanyDeletedDto"
                  - $ref: "#/components/schemas/FullMfaRequiredDto"
                  - $ref: "#/components/schemas/TokenTwoFaRequiredDto"
                  - $ref: "#/components/schemas/PhoneMfaRequiredDto"
        "400":
          description: In case if either user provided the wrong password or provided
            wrong recovery keys, the bad request error will be thrown.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/WrongCredentialsException"
                  - $ref: "#/components/schemas/WrongRecoveryKeysException"
      tags: *a4
      security:
        - x-access-token: []
        - basicAuth: []
...
info:
  title: Mnemosyne
  description: Documentation of Mnemosyne API.
  version: 0.0.1
  contact: {}
...
servers: []
components:
  securitySchemes:
    x-access-token:
      scheme: bearer
      bearerFormat: JWT
      description: "Please enter the JWT token in the following format: Bearer <JWT>"
      name: x-access-token
      type: apiKey
      in: header
    basicAuth:
      type: http
      scheme: basic
      description: Basic API authentication username and password
      in: header
  schemas:
...
    UserNotFoundException:
      type: object
      properties:
        message:
          type: string
          description: Marketplace user not found error message
          example: user-not-found
      required:
        - message
...
    FullMfaRequiredDto:
      type: object
      properties:
        message:
          type: string
          description: Full MFA (token + phone) required response message
          example: full-mfa-required
      required:
        - message
    TokenTwoFaRequiredDto:
      type: object
      properties:
        message:
          type: string
          description: Token MFA required response message
          example: token-two-fa-required
      required:
        - message
    PhoneMfaRequiredDto:
      type: object
      properties:
        message:
          type: string
          description: Phone MFA required response message
          example: phone-required
      required:
        - message
...
    SmsExpiredException:
      type: object
      properties:
        message:
          type: string
          description: SMS expired error message
          example: sms-expired
      required:
        - message
    WrongCredentialsException:
      type: object
      properties:
        message:
          type: string
          description: Wrong credentials error message
          example: wrong-credentials
      required:
        - message
    WrongCodeException:
      type: object
      properties:
        message:
          type: string
          description: Wrong code error message
          example: wrong-code
      required:
        - message
...
    TacNotAcceptedException:
      type: object
      properties:
        message:
          type: string
          description: T&C not accepted error message
          example: tac-not-accepted
      required:
        - message
...
    CreateCompanyDto:
      type: object
      properties:
        companyLocation:
          type: string
          description: Company location
          example: Cupertino, CA 95014
        companyName:
          type: string
          description: Company name
          example: Apple Inc.
        companyWebsite:
          type: string
          description: Company website
          example: apple.com
        companyOwnerEmail:
          type: string
          description: Email address
          example: joe@doe.com
        companyMembers:
          description: List of company members emails
          example:
            - email: tim.cook@icloud.com
              role: DEFAULT
          type: array
          items:
            type: array
        tac:
          type: boolean
          description: Terms & Conditions
          example: "true"
        language:
          type: string
          enum:
            - ru
            - pl
            - en
          description: One of 3 available languages
          example: pl
      required:
        - companyLocation
        - companyName
        - companyWebsite
        - companyOwnerEmail
        - tac
    CompanyCreatedDto:
      type: object
      properties:
        message:
          type: string
          description: "Company created response message "
          example: company-created
      required:
        - message
    CompanyExistsException:
      type: object
      properties:
        message:
          type: string
          description: Company already exists error message
          example: company-already-exists
      required:
        - message
    GetCompanyByIdDto:
      type: object
      properties:
        companyName:
          type: string
          description: Company name
          example: Apple Inc.
        companyLocation:
          type: string
          description: Company location
          example: Cupertino, CA 95014
        companyWebsite:
          type: string
          description: Company website
          example: apple.com
        companyOwnerId:
          type: string
          description: User ID
          example: 4eea51f7-ebf1-4470-b54e-eb3b8c79bf2c
        companyOwnerEmail:
          type: string
          description: Email address
          example: joe@doe.com
      required:
        - companyName
        - companyLocation
        - companyWebsite
        - companyOwnerId
        - companyOwnerEmail
    ParseException:
      type: object
      properties:
        message:
          type: string
          description: Parse exception
          example: parse-error
      required:
        - message
    CompanyNotFoundException:
      type: object
      properties:
        message:
          type: string
          description: Company not found error message
          example: company-not-found
      required:
        - message
...
    GetCompanyUsersDto:
      type: object
      properties:
        companyUsers:
          description: List of company users (id and email)
          example:
            - id: 4eea51f7-ebf1-4470-b54e-eb3b8c79bf2c
              email: joe@doe.com
              role:
                id: 4eea51f7-ebf1-4470-b54e-eb3b8c79bf8c
                name: PRIMARY_ADMIN
              registrationHash:
                confirmed: "false"
                createdAt: 2023-11-17T11:42:26.534Z
          type: array
          items:
            type: array
        count:
          type: number
          description: Count for the pagination
          example: "5"
      required:
        - companyUsers
        - count
    CompanyUpdatedDto:
      type: object
      properties:
        message:
          type: string
          description: Company update response message
          example: company-updated
      required:
        - message
    UpdateCompanyDto:
      type: object
      properties:
        companyLocation:
          type: string
          description: Company location
          example: Cupertino, CA 95014
        companyName:
          type: string
          description: Company name
          example: Apple Inc.
        companyWebsite:
          type: string
          description: Company website
          example: apple.com
      required:
        - companyLocation
        - companyName
        - companyWebsite
    TransferOwnershipDto:
      type: object
      properties:
        newCompanyOwnerEmail:
          type: string
          description: Email address
          example: joe@doe.com
        newRoleForOldOwnerId:
          type: string
          description: Company role ID
          example: 4eea51f7-ebf1-4470-b54e-eb3b8c79bf2c
        mfaCode:
          type: string
          description: User authentication application code - MFA
          example: "123123"
        phoneCode:
          type: string
          description: User phone code - MFA
          example: "123123"
        language:
          type: string
          enum:
            - ru
            - pl
            - en
          description: One of 3 available languages
          example: pl
      required:
        - newCompanyOwnerEmail
        - newRoleForOldOwnerId
...
    CompanyOwnershipTransferredDto:
      type: object
      properties:
        message:
          type: string
          description: Company ownership transferred response message
          example: company-ownership-transferred
      required:
        - message
    DeleteCompanyDto:
      type: object
      properties:
        password:
          type: string
          description: User password
          example: 78ui&*UI
        passphrase:
          type: string
          description: Passphrase
          example: "123123123"
        recoveryKeys:
          description: List of recovery keys
          example:
            - e54ee7e285fbb....*length 1024*
            - e54ee7e285fbb....*length 1024*
            - e54ee7e285fbb....*length 1024*
            - e54ee7e285fbb....*length 1024*
            - e54ee7e285fbb....*length 1024*
          type: array
          items:
            type: array
            maxLength: 5
            minLength: 5
        mfaCode:
          type: string
          description: User authentication application code - MFA
          example: "123123"
        phoneCode:
          type: string
          description: User authentication application code - MFA
          example: "123123"
        language:
          type: string
          enum:
            - ru
            - pl
            - en
          description: One of 3 available languages
          example: pl
      required:
        - password
        - passphrase
        - recoveryKeys
    CompanyDeletedDto:
      type: object
      properties:
        message:
          type: string
          description: Company deleted response message
          example: company-deleted
      required:
        - message
    WrongRecoveryKeysException:
      type: object
      properties:
        message:
          type: string
          description: Wrong recovery keys error message
          example: wrong-recovery-keys
      required:
        - message
...